pyproject.toml
[project]
name = "coderag-store"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "bs4>=0.0.2",
    "daytona>=0.117.0",
    "daytona-sdk>=0.117.0",
    "ddgs>=9.9.1",
    "e2b-code-interpreter>=2.3.0",
    "iprogress>=0.4",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.7",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "langgraph-api>=0.5.23",
    "langgraph-cli[inmem]>=0.4.7",
    "matplotlib>=3.10.7",
    "modal>=1.2.2",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pip>=25.3",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentry-sdk>=2.43.0",
    "tqdm>=4.67.1",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
]
